+++ Sources/BaseFoundationiOS/Common/Extensions/ColorExtension.swift
@@ -14,4 +14,17 @@ extension Color {
         let blue = Double((hex & 0xff) >> 0) / 255.0
         self.init(.sRGB, red: red, green: green, blue: blue, opacity: opacity)
     }
+    
+    public func toHexString() -> String {
+        let components = UIColor(self).cgColor.components
+        let r: CGFloat = components?[0] ?? 0.0
+        let g: CGFloat = components?[1] ?? 0.0
+        let b: CGFloat = components?[2] ?? 0.0
+
+        let hexString = String.init(format: "#%02lX%02lX%02lX",
+                                    lroundf(Float(r * 255)),
+                                    lroundf(Float(g * 255)),
+                                    lroundf(Float(b * 255)))
+        return hexString
+    }
 }